requirements.txt
aiofiles==25.1.0
aiogram==3.24.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-types==0.7.0
appdirs==1.4.4
async-generator==1.10
attrs==25.4.0
beautifulsoup4==4.14.2
bs4==0.0.2
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
colorama==0.4.6
cryptography==46.0.4
Cython==3.2.2
frozenlist==1.8.0
google-ai-generativelanguage==0.6.15
google-api-core==2.29.0
google-api-python-client==2.189.0
google-auth==2.49.0.dev0
google-auth-httplib2==0.3.0
google-generativeai==0.8.6
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
hidapi==0.14.0.post4
httplib2==0.31.2
idna==3.11
importlib_metadata==8.7.1
intelhex==2.3.0
magic-filter==1.0.12
markdown-it-py==4.0.0
mdurl==0.1.2
multidict==6.7.1
mypy_extensions==1.1.0
outcome==1.3.0.post0
packaging==26.0
parse_pip_search==0.0.1
propcache==0.4.1
proto-plus==1.27.1
protobuf==5.29.6
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pyedbglib==2.24.2.18
Pygments==2.19.2
pymcuprog==3.19.4.61
pyparsing==3.3.2
pyserial==3.5
PySocks==1.7.1
python-dotenv==1.2.1
PyYAML==6.0.3
requests==2.32.5
rich==14.2.0
selenium==4.40.0
setuptools==80.9.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8
tqdm==4.67.3
trio==0.32.0
trio-typing==0.10.0
trio-websocket==0.12.2
types-certifi==2021.10.8.3
types-urllib3==1.26.25.14
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.3
websocket-client==1.9.0
wsproto==1.3.2
yarl==1.22.0
zipp==3.23.0
